Rename: query, mutation 함수에 맞게 폴더 변경
