+++ src/hooks/queries/useGetLectureInfo.ts
